+++ projects/_03_independent_cqc/_01_merge/jobs/merge_ind_cqc_data.py
@@ -4,6 +4,7 @@ import sys
 os.environ["SPARK_VERSION"] = "3.5"
 from typing import Optional
 from pyspark.sql import DataFrame
+
 import utils.cleaning_utils as cUtils
 from utils import utils
 from utils.column_names.capacity_tracker_columns import (
@@ -24,9 +25,6 @@ from utils.column_names.cleaned_data_files.cqc_pir_cleaned import (
 from utils.column_names.cleaned_data_files.ons_cleaned import (
     OnsCleanedColumns as ONSClean,
 )
-from utils.column_names.ind_cqc_pipeline_columns import (
-    DimensionPartitionKeys as DimensionKeys,
-)
 from utils.column_names.ind_cqc_pipeline_columns import PartitionKeys as Keys
 from utils.column_values.categorical_column_values import (
     LocationType,
@@ -40,48 +38,25 @@ cleaned_cqc_locations_columns_to_import = [
     CQCLClean.cqc_location_import_date,
     CQCLClean.location_id,
     CQCLClean.name,
+    CQCLClean.postal_code,
     CQCLClean.provider_id,
     CQCLClean.cqc_sector,
-    CQCLClean.type,
     CQCLClean.registration_status,
     CQCLClean.imputed_registration_date,
     CQCLClean.dormancy,
+    CQCLClean.care_home,
     CQCLClean.number_of_beds,
-    CQCLClean.related_location,
-    Keys.year,
-    Keys.month,
-    Keys.day,
-    Keys.import_date,
-]
-gac_dim_columns_to_import = [
-    CQCLClean.location_id,
+    CQCLClean.imputed_regulated_activities,
     CQCLClean.imputed_gac_service_types,
     CQCLClean.services_offered,
-    CQCLClean.primary_service_type,
-    CQCLClean.care_home,
-    Keys.import_date,
-    DimensionKeys.last_updated,
-]
-ra_dim_columns_to_import = [
-    CQCLClean.location_id,
-    CQCLClean.imputed_regulated_activities,
-    CQCLClean.registered_manager_names,
-    Keys.import_date,
-    DimensionKeys.last_updated,
-]
-specialism_dim_columns_to_import = [
-    CQCLClean.location_id,
     CQCLClean.imputed_specialisms,
     CQCLClean.specialisms_offered,
     CQCLClean.specialist_generalist_other_dementia,
     CQCLClean.specialist_generalist_other_lda,
     CQCLClean.specialist_generalist_other_mh,
-    Keys.import_date,
-    DimensionKeys.last_updated,
-]
-pcm_dim_columns_to_import = [
-    CQCLClean.location_id,
-    CQCLClean.postal_code,
+    CQCLClean.related_location,
+    CQCLClean.primary_service_type,
+    CQCLClean.registered_manager_names,
     ONSClean.contemporary_ons_import_date,
     ONSClean.contemporary_cssr,
     ONSClean.contemporary_region,
@@ -95,8 +70,10 @@ pcm_dim_columns_to_import = [
     ONSClean.current_region,
     ONSClean.current_lsoa21,
     ONSClean.current_rural_urban_ind_11,
+    Keys.year,
+    Keys.month,
+    Keys.day,
     Keys.import_date,
-    DimensionKeys.last_updated,
 ]
 cleaned_ascwds_workplace_columns_to_import = [
     AWPClean.ascwds_workplace_import_date,
@@ -128,10 +105,6 @@ cleaned_ct_care_home_columns_to_import = [
 
 def main(
     cleaned_cqc_location_source: str,
-    gac_dim_source: str,
-    reg_act_dim_source: str,
-    spec_dim_source: str,
-    pcm_dim_source: str,
     cleaned_cqc_pir_source: str,
     cleaned_ascwds_workplace_source: str,
     cleaned_ct_non_res_source: str,
@@ -141,48 +114,9 @@ def main(
     spark = utils.get_spark()
     spark.sql("set spark.sql.broadcastTimeout = 2000")
 
-    # TODO - Polars conversion - apply .fliter on scan_parquet instead of reading full dataframe and then filtering
     cqc_location_df = utils.read_from_parquet(
         cleaned_cqc_location_source,
-        selected_columns=cleaned_cqc_locations_columns_to_import,
-    )
-    cqc_filtered_df = utils.select_rows_with_value(
-        cqc_location_df, CQCLClean.type, LocationType.social_care_identifier
-    )
-    cqc_filtered_df = utils.select_rows_with_value(
-        cqc_filtered_df, CQCLClean.registration_status, RegistrationStatus.registered
-    )
-    cqc_filtered_df = utils.select_rows_with_value(
-        cqc_filtered_df, CQCLClean.cqc_sector, Sector.independent
-    )
-
-    gac_dim_df = utils.read_from_parquet(
-        gac_dim_source,
-        selected_columns=gac_dim_columns_to_import,
-    )
-    cqc_filtered_df = utils.join_dimension(
-        cqc_filtered_df, gac_dim_df, CQCLClean.location_id
-    )
-
-    reg_act_dim_df = utils.read_from_parquet(
-        reg_act_dim_source, selected_columns=ra_dim_columns_to_import
-    )
-    cqc_filtered_df = utils.join_dimension(
-        cqc_filtered_df, reg_act_dim_df, CQCLClean.location_id
-    )
-
-    spec_dim_df = utils.read_from_parquet(
-        spec_dim_source, selected_columns=specialism_dim_columns_to_import
-    )
-    cqc_filtered_df = utils.join_dimension(
-        cqc_filtered_df, spec_dim_df, CQCLClean.location_id
-    )
-
-    pcm_dim_df = utils.read_from_parquet(
-        pcm_dim_source, selected_columns=pcm_dim_columns_to_import
-    )
-    cqc_filtered_df = utils.join_dimension(
-        cqc_filtered_df, pcm_dim_df, CQCLClean.location_id
+        # selected_columns=cleaned_cqc_locations_columns_to_import,
     )
 
     ascwds_workplace_df = utils.read_from_parquet(
@@ -200,31 +134,35 @@ def main(
         selected_columns=cleaned_ct_care_home_columns_to_import,
     )
 
-    cqc_filtered_df = join_data_into_cqc_df(
-        cqc_filtered_df,
+    ind_cqc_location_df = utils.select_rows_with_value(
+        cqc_location_df, CQCLClean.cqc_sector, Sector.independent
+    )
+
+    ind_cqc_location_df = join_data_into_cqc_df(
+        ind_cqc_location_df,
         cqc_pir_df,
         CQCPIRClean.location_id,
         CQCPIRClean.cqc_pir_import_date,
         CQCPIRClean.care_home,
     )
 
-    cqc_filtered_df = join_data_into_cqc_df(
-        cqc_filtered_df,
+    ind_cqc_location_df = join_data_into_cqc_df(
+        ind_cqc_location_df,
         ascwds_workplace_df,
         AWPClean.location_id,
         AWPClean.ascwds_workplace_import_date,
     )
 
-    cqc_filtered_df = join_data_into_cqc_df(
-        cqc_filtered_df,
+    ind_cqc_location_df = join_data_into_cqc_df(
+        ind_cqc_location_df,
         ct_non_res_df,
         CTNRClean.cqc_id,
         CTNRClean.ct_non_res_import_date,
         CTNRClean.care_home,
     )
 
-    cqc_filtered_df = join_data_into_cqc_df(
-        cqc_filtered_df,
+    ind_cqc_location_df = join_data_into_cqc_df(
+        ind_cqc_location_df,
         ct_care_home_df,
         CTCHClean.cqc_id,
         CTCHClean.ct_care_home_import_date,
@@ -232,7 +170,7 @@ def main(
     )
 
     utils.write_to_parquet(
-        cqc_filtered_df,
+        ind_cqc_location_df,
         destination,
         mode="overwrite",
         partitionKeys=PartitionKeys,
@@ -283,10 +221,6 @@ if __name__ == "__main__":
     print(f"Job parameters: {sys.argv}")
     (
         cleaned_cqc_location_source,
-        gac_dim_source,
-        reg_act_dim_source,
-        spec_dim_source,
-        pcm_dim_source,
         cleaned_cqc_pir_source,
         cleaned_ascwds_workplace_source,
         cleaned_ct_non_res_source,
@@ -296,22 +230,6 @@ if __name__ == "__main__":
         (
             "--cleaned_cqc_location_source",
             "Source s3 directory for parquet CQC locations cleaned dataset",
-        ),
-        (
-            "--gac_services_dimension_source",
-            "Source s3 directory for parquet GAC services dimension",
-        ),
-        (
-            "--regulated_activities_dimension_source",
-            "Source s3 directory for parquet Regulated Activities dimension",
-        ),
-        (
-            "--specialisms_dimension_source",
-            "Source s3 directory for parquet Services dimension",
-        ),
-        (
-            "--postcode_matching_dimension_source",
-            "Source s3 directory for parquet Postcode Matching dimension",
         ),
         (
             "--cleaned_cqc_pir_source",
@@ -336,10 +254,6 @@ if __name__ == "__main__":
     )
     main(
         cleaned_cqc_location_source,
-        gac_dim_source,
-        reg_act_dim_source,
-        spec_dim_source,
-        pcm_dim_source,
         cleaned_cqc_pir_source,
         cleaned_ascwds_workplace_source,
         cleaned_ct_non_res_source,